refactor: removing dead code
