+++ main/src/domains/shared/components/app-router.tsx
@@ -14,9 +14,6 @@ import { TableScreen } from "../../tables/screens/table";
 import { ErrorScreen } from "../screens/error";
 import { Testing } from "../testing-temp/testing";
 import { Chat } from "../../chat";
-import { RoomsScreen } from "../../tables/screens/rooms";
-import { RoomScreen } from "../../tables/screens/room";
-import { GameScreen } from "../../games/screens/game";
 import { DashboardScreen } from "../../dashboard/screens/dashboard";
 import { UISystemScreen } from "@/domains/ui-system/screen";
 import { ChannelScreen } from "@/domains/channel/screens/channel";
